requirements.txt
selenium
pytest
pytest-selenium
pytest-html
PyYAML
